requirements.txt
numpy
scipy
lxml
matplotlib
networkx
codecov
pygraphviz
pytest-cov
pytest-subtests
pyro-ppl==1.5.1
torchviz
torch==1.7.1
torch_geometric==1.7.0
pytorch3d==0.3
meshcat
